Allow users to disable temp dir file logging.
If your on Windows or some other OS where for some reason its not a ram disk then you may want this but disabling this means if any bug happens debugging it might be harder.

This setting will be disabled by default once borz is known to be pretty stable.