Make simulate actually work.
Before it wouldn't print or try to compile since it was still checking dates on files which shouldn't mater for simulate so theres now a custom function in CppBuilder that will just give DateTime.Now.

Maybe in the future there should be a config to restore the file checks if you want them.